requirements.txt
faker==33.3.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
